Many test parameters in one object
Almost all parameters of the test are collected in one object, except the check-in and check-out dates.
The test file has been renamed.
